requirements.txt
Django==3.1
djangorestframework==3.12.2
drf-yasg==1.20.0
requests==2.25.1
sqlparse==0.4.1
pytest==5.3.4
pytest-django==3.7.0
six==1.15.0
django-filter==2.4.0